+++ doc/version_history.md
@@ -1,4 +1,7 @@
 ## 版本更新记录
+##### v6.x
+> - [Upgrade] 优化 RestControllerAspect 切面日志处理
+> - [Upgrade] testspringboot 增加异步请求接口demo
 ##### v6.0.0
 > - [Upgrade] 删除 acp-spring-boot-starter-ws 和 acp-webservice 模块，彻底废弃 webservice 相关功能的支持
 > - [Upgrade] java 1.8 环境中删除 testscala 模块